Split InternalMath
